requirements.txt



flask==3.1.0
llama-cpp-python==0.3.16

